requirements.txt
Faker==13.14.0
PyYAML==6.0